% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.windrose.R
\name{print}
\alias{print}
\alias{print.windrose}
\title{Prints a windrose object to console}
\usage{
print(data_rose)
}
\arguments{
\item{data}{(required) a windrose object}
}
\description{
Prints a windrose object to the console.
}
