% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.windrose.R
\name{summary}
\alias{summary}
\alias{summary.windrose}
\title{Prints a summary of a windrose object}
\usage{
summary(data_rose)
}
\arguments{
\item{data}{(required) a windrose object}
}
\value{
returns a summary string for a windrose object.
}
\description{
Creates and prints a summary of a windrose object.
}
\seealso{
\code{brewer.pal}.
}
