% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windrose.R
\name{windrose}
\alias{windrose}
\title{Create a windrose object}
\usage{
data_rose <- windrose(data, speed_column_name, direction_column_name)
}
\arguments{
\item{data}{(optional) a data.frame containing, at least, columns for wind speed and wind direction}

\item{spd}{(required) either a vector or the data frame column name containing wind speed data. Do not enclose in quotes.}

\item{dir}{(required) either a vector or the data frame column name containing wind direction data. Do not enclose in quotes.}

\item{spdmin}{(optional) numeric. the minimum wind speed to plot}

\item{spdmax}{(optional) numeric. the maximum wind speed to plot}

\item{spdseq}{(optional) numeric vector. the cut points used for wind speed color fills. must have at least five elements.}

\item{palette}{(optional) one of the RColorBrewer ColorBrewer palettes. Can be changed when plotting.}

\item{countmax}{(optional) numeric. the maximum number of counts for wind direction data in each bin.}

\item{plot_rose}{(optional) boolean. if \code{TRUE}, plots the rose graph when generating the windrose object.}
}
\value{
a windrose object.
}
\description{
Create a windrose object, mainly used for plotting rose plots that display both wind direction and speed data.
}
\seealso{
\code{brewer.pal}.
}
