% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.windrose.R
\name{plot}
\alias{plot}
\alias{plot.windrose}
\title{Plots a waverose or windrose object using ggplot2}
\usage{
\method{plot}{windrose}(data, palette = "YlGnBu",
  plot_theme = "theme_minimal", t_legend = "Wind Speed (m/s)",
  frequency_relative = T, blanked = F, ...)
}
\arguments{
\item{data}{data object of class windrose}

\item{palette}{RColorBrewer palettes}

\item{plot_theme}{character string specifying a ggplot theme, e.g. "theme_minimal". Also supports ggthemes and xkcd.}

\item{t_legend}{character Text of legend.}

\item{frequency_relative}{logical if the relative frequency of wind velocity classes is computed. Default is TRUE.}

\item{blanked}{logical if windrose annotations are eliminated. Default is FALSE.}

\item{no_legend}{logical If legend is eliminated. Default is FALSE.}

}
\value{
returns a ggplot2 graph object
}
\seealso{
\code{ggplot2}, \code{ggtheme} and \code{brewer.pal}.
}
