% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windrose.R
\name{windrose}
\alias{windrose}
\title{Create a windrose object}
\usage{
data_rose <- windrose(data, speed_column_name, direction_column_name)
}
\arguments{
\item{data}{data.frame Data.frame cointaining data columns for wind speed and wind direction. Optional.}

\item{spd}{object Data.frame column name containing wind speed data. Do not enclose in quotes. Required.}

\item{dir}{object Data.frame column na containing wind direction data. Do not enclose in quotes. Required.}

\item{spdmin}{numeric Minimum wind speed. Optional.}

\item{spdmax}{numeric Maximum wind speed. Optional.}

\item{spdseq}{vector Cut points used for wind speed color fills. They must have at least five elements. Optional.}

\item{spdcalm}{numeric Minimum wind speed treshshold for calm wind condition. Default is 0.3.}

\item{palette}{character RColorBrewer palette. Can be changed when plotting. Optional.}

\item{countmax}{numeric The maximum number of counts for wind direction data in each bin. Optional.}

\item{calm}{boolean If \code{TRUE}  the calm data  is given. Optional.}

\item{plot_rose}{boolean If \code{TRUE}  the rose graph is given. Optional.}
}
\value{
list A windrose object.
}
\description{
Create a windrose object, mainly used for plotting rose plots that display both wind direction and speed data.
}
\seealso{
\code{brewer.pal}.
}
